+++ src/anyio/__init__.py
@@ -12,7 +12,6 @@ __all__ = (
     'ClosedResourceError',
     'DelimiterNotFound',
     'EndOfStream',
-    'ExceptionGroup',
     'IncompleteRead',
     'TypedAttributeLookupError',
     'WouldBlock',
@@ -68,8 +67,7 @@ from ._core._eventloop import (
     sleep_until)
 from ._core._exceptions import (
     BrokenResourceError, BrokenWorkerProcess, BusyResourceError, ClosedResourceError,
-    DelimiterNotFound, EndOfStream, ExceptionGroup, IncompleteRead, TypedAttributeLookupError,
-    WouldBlock)
+    DelimiterNotFound, EndOfStream, IncompleteRead, TypedAttributeLookupError, WouldBlock)
 from ._core._fileio import AsyncFile, Path, open_file, wrap_file
 from ._core._resources import aclose_forcefully
 from ._core._signals import open_signal_receiver